erepo: move get_rev to repo
